refactor and simplify badges
* cleanup WIP code
