+++ meal-ui/src/styles/local/partials/1-elements/1-elements.scss
@@ -16,3 +16,6 @@
 
 /* forms */
 @import 'forms/forms.scss';
+
+/* element collections */
+@import 'tiled-badges.scss';
