+++ meal-ui/src/styles/local/partials/1-elements/tiled-badges.scss
@@ -1,0 +1,12 @@
+@use '@carbon/react/scss/spacing' as *;
+
+.tiled-badges {
+	display: flex;
+	flex-direction: column;
+	gap: $spacing-02;
+
+	.tiled-badges__row {
+		display: flex;
+		gap: $spacing-02;
+	}
+}
