+++ meal-ui/src/styles/local/partials/2-components/horizontal-card/horizontal-card.scss
@@ -31,10 +31,5 @@
 		h4 {
 			@include text-header();
 		}
-
-		.content-slot__badges {
-			display: flex;
-			gap: 8px;
-		}
 	}
 }
